--- SeleniumWebdriver_Udemy/src/main/resources/frame.html
@@ -1,1 +1,0 @@
-<input type="button" id="frameButton" value="Dentro do Frame" onclick="alert('Frame OK!')" />
