--- SeleniumWebdriver_Udemy/src/test/java/TesteAlert.java
@@ -1,65 +1,0 @@
-import org.junit.After;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
-import org.openqa.selenium.Alert;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-
-public class TesteAlert {
-    WebDriver driver;
-
-    @Before
-    public void preCondicao() {
-        System.setProperty("webdriver.chrome.driver","C:\\Users\\isabellaSzrajbman\\Documents\\MeusProjetos\\SeleniumWebdriver_Udemy\\drivers\\chromedriver97\\chromedriver.exe");
-        driver = new ChromeDriver();
-        driver.navigate().to("file:///" + System.getProperty("user.dir") + "/src/main/resources/componentes.html");
-        driver.manage().window().maximize(); // maximiza a janela do navegador.
-    }
-
-    @Test
-    public void deveInteragirComAlertSimples() {
-        driver.findElement(By.id("alert")).click();
-        Alert alert = driver.switchTo().alert();
-        String texto = alert.getText();
-        Assert.assertEquals("Alert Simples", texto);
-        alert.accept();
-        driver.findElement(By.id("elementosForm:nome")).sendKeys(texto);
-    }
-
-    @Test
-    public void deveInteragirComAlertConfirm(){
-        driver.findElement(By.id("confirm")).click();
-        Alert alerta = driver.switchTo().alert();
-        Assert.assertEquals("Confirm Simples", alerta.getText());
-        alerta.accept();
-        Assert.assertEquals("Confirmado", alerta.getText());
-        alerta.accept();
-
-        driver.findElement(By.id("confirm")).click();
-        alerta = driver.switchTo().alert();
-        Assert.assertEquals("Confirm Simples", alerta.getText());
-        alerta.dismiss();
-        Assert.assertEquals("Negado", alerta.getText());
-        alerta.dismiss();
-    }
-
-    @Test
-    public void deveInteragirComAlertPrompt(){
-        driver.findElement(By.id("prompt")).click();
-        Alert alerta = driver.switchTo().alert();
-        Assert.assertEquals("Digite um numero", alerta.getText());
-        alerta.sendKeys("12");
-        alerta.accept();
-        Assert.assertEquals("Era 12?", alerta.getText());
-        alerta.accept();
-        Assert.assertEquals(":D", alerta.getText());
-        alerta.accept();
-    }
-
-    @After
-    public void posCondicao() {
-        //driver.quit();
-    }
-}
