--- SeleniumWebdriver_Udemy/src/test/java/TesteFramesEJanelas.java
@@ -1,59 +1,0 @@
-import org.junit.After;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
-import org.openqa.selenium.Alert;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-
-public class TesteFramesEJanelas {
-    WebDriver driver;
-
-    @Before
-    public void preCondicao() {
-        System.setProperty("webdriver.chrome.driver","C:\\Users\\isabellaSzrajbman\\Documents\\MeusProjetos\\SeleniumWebdriver_Udemy\\drivers\\chromedriver97\\chromedriver.exe");
-        driver = new ChromeDriver();
-        driver.navigate().to("file:///" + System.getProperty("user.dir") + "/src/main/resources/componentes.html");
-        driver.manage().window().maximize(); // maximiza a janela do navegador.
-    }
-
-    @Test
-    public void deveInteragirComFrames(){
-        driver.switchTo().frame("frame1");
-        driver.findElement(By.id("frameButton")).click();
-        Alert alert = driver.switchTo().alert();
-        String msg = alert.getText();
-        Assert.assertEquals("Frame OK!", msg);
-        alert.accept();
-
-        driver.switchTo().defaultContent();
-        driver.findElement(By.id("elementosForm:nome")).sendKeys(msg);
-    }
-
-    @Test
-    public void deveInteragirComJanelas(){
-        driver.findElement(By.id("buttonPopUpEasy")).click();
-        driver.switchTo().window("Popup");
-        driver.findElement(By.tagName("textarea")).sendKeys("Deu certo?");
-        driver.close();
-        driver.switchTo().window("");
-        driver.findElement(By.tagName("textarea")).sendKeys("e agora?");
-    }
-
-    @Test
-    public void deveInteragirComJanelasSemTitulo(){
-        driver.findElement(By.id("buttonPopUpHard")).click();
-        System.out.println(driver.getWindowHandle());
-        System.out.println(driver.getWindowHandles());
-        driver.switchTo().window((String) driver.getWindowHandles().toArray()[1]);
-        driver.findElement(By.tagName("textarea")).sendKeys("Deu certo?");
-        driver.switchTo().window((String) driver.getWindowHandles().toArray()[0]);
-        driver.findElement(By.tagName("textarea")).sendKeys("E agora?");
-    }
-
-    @After
-    public void posCondicao() {
-        driver.quit();
-    }
-}
